structure changes. shadow still not good
