+++ grid_render/src/custom_elements/mod.rs
@@ -1,7 +1,9 @@
 mod custom_element_definition;
 pub(crate) mod bq_table_custom_element;
-pub(crate) mod table_plot;
+pub(crate) mod data_table_element;
 mod bq_to_table;
+mod data_table_controls_element;
+pub(crate) mod data_table_shadow_element;
 
 use self::custom_element_definition::CustomElementDefinition;
 use std::{error::Error, fmt, str::FromStr};
